pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ppa-for-industries"
version = "0.0.1"
description = "This is a Tool for the Open-Energy-Data-Server which analyzes PPAs for various industries in Germany."
authors = [{ name = "Jonathan Sejdija", email = "sejdija@fh-aachen.de"}, { name = "Florian Maurer", email = "maurer@fh-aachen.de"}]
license = {text = "AGPL-3.0-or-later"}
readme = "README.md"
keywords = ["PPAs"]

classifiers=[
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "pandas >=2.0.0",
    "sqlalchemy",
    "psycopg2-binary",
    "numpy",
    "python-dotenv",
    "babel",
    "plz2nuts"
]

[project.urls]
Repository = "https://github.com/NOWUM/ppa-for-industries"
Issues = "https://github.com/NOWUM/ppa-for-industries/issues"


[tool.setuptools]
packages = {find = {}}

[tool.ruff]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "G", "PIE"]
ignore = ["E501", "G004", "E731"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "I001", # allow unsorted imports in __init__.py
    "F401", # allow unused imports in __init__.py
]
